requirements.txt
opencv-python-headless
pyyaml
tqdm
numpy
easydict==1.9.0
scikit-image==0.17.2
tensorflow
joblib
matplotlib
pandas
albumentations==0.5.2
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
scikit-learn==0.24.2
wldhx.yadisk-direct
black==22.3.0
cloudpickle
future
fvcore
omegaconf
pillow==8.4.0
pycocotools
seaborn
